G004-remove_deprecated-ModelicaSystem_rewrite_set_functions2
[ModelicaSystemABC] remove code for (depreciated) arguments in set*() methods

* define code in the compatibility layer in class ModelicaSystem

[test_ModelicaSystem(OMC)] update tests

* for new version: remove usage of old definition
* for compatibility version: test old definition
